docs/requirements.txt
click
pyyaml
Sphinx>=2.0
sphinx-autodoc-typehints
sphinx-rtd-theme
sphinx-click
